updated the way student schedules are fetched and added ScheduleDTO.java to wrap student schedule into an object
