+++ src/main/resources/create_dbs.sql
@@ -115,7 +115,7 @@ CREATE TABLE COURSE_GRADE(
     STUDENT_ID BIGINT NOT NULL,
     COURSE_ID BIGINT NOT NULL,
     MODIFIED_AT TIMESTAMP,
-    GRADE VARCHAR(2),
+    GRADE VARCHAR(6),
     FOREIGN KEY (STUDENT_ID) REFERENCES STUDENT(ID),
     FOREIGN KEY (COURSE_ID) REFERENCES COURSE(ID),
     CONSTRAINT COURSE_GRADE_KEY PRIMARY KEY (STUDENT_ID, COURSE_ID)
