+++ src/main/resources/init_course_grades.sql
@@ -1,0 +1,6 @@
+INSERT INTO COURSE_GRADE(STUDENT_ID, COURSE_ID, GRADE) VALUES (14, 2, 'AMinus');
+INSERT INTO COURSE_GRADE(STUDENT_ID, COURSE_ID, GRADE) VALUES (15, 2, 'B');
+INSERT INTO COURSE_GRADE(STUDENT_ID, COURSE_ID, GRADE) VALUES (16, 2, 'C');
+INSERT INTO COURSE_GRADE(STUDENT_ID, COURSE_ID, GRADE) VALUES (17, 2, 'DMinus');
+INSERT INTO COURSE_GRADE(STUDENT_ID, COURSE_ID, GRADE) VALUES (18, 2, 'E');
+INSERT INTO COURSE_GRADE(STUDENT_ID, COURSE_ID, GRADE) VALUES (19, 2, 'A');
