--- src/test/java/PersonTest/PersonTest.java
@@ -1,4 +1,0 @@
-package PersonTest;
-
-public class PersonTest {
-}
