+++ src/test/java/com/example/ums/enumstests/PersonTest/PersonTest.java
@@ -1,0 +1,4 @@
+package com.example.ums.enumstests.PersonTest;
+
+public class PersonTest {
+}
